changes after the latest rebase with master
