+++ internal/cmd/operator-sdk/generate/internal/genutil.go
@@ -21,10 +21,14 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/util/validation"
 
 	"github.com/blang/semver"
 	apiextv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apiextv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
+	"sigs.k8s.io/kubebuilder/pkg/model/config"
 	"sigs.k8s.io/yaml"
 )
 
@@ -172,4 +176,25 @@ func IsNotExist(path string) bool {
 	}
 	_, err := os.Stat(path)
 	return err != nil && errors.Is(err, os.ErrNotExist)
+}
+
+// GetOperatorName returns the name of the operator which is by default the projectName attribute of the PROJECT file
+// However, the Go projects built with the plugin version v2 has not this attribute and then, for this case
+// the operatorName will be the current directory.
+func GetOperatorName(cfg *config.Config) (string, error) {
+	if cfg.ProjectName != "" {
+		return cfg.ProjectName, nil
+	}
+	if cfg.IsV3() {
+		return "", errors.New("project config file must contain 'projectName'")
+	}
+	dir, err := os.Getwd()
+	if err != nil {
+		return "", fmt.Errorf("error getting current directory: %v", err)
+	}
+	projectName := strings.ToLower(filepath.Base(dir))
+	if err := validation.IsDNS1123Label(projectName); err != nil {
+		return "", fmt.Errorf("project name (%s) is invalid: %v", projectName, err)
+	}
+	return projectName, nil
 }
